DAOS-12447 vos: fix LRU array free lru_sub list corruption
master-commit: f45b589bd0467e2fa6d8a11ed4068c32ecbcfa36

When there is no free entry in a lru_sub, it is expected to be removed
from the free list of the LRU array. But under some cases, such as DTX
reindex that may trigger lrua_remove_entry, it does not do that, as to
related lru_sub may be added into the LRU array free lru_sub list more
that once. Then when traverse such list, it may fall into dead loop.

Signed-off-by: Fan Yong <fan.yong@intel.com>