DAOS-12689 control: Don't return an error if dlsym succeeded (#11596)
dlerror() may not be thread-safe, depending on the implementation.
It is possible that after clearing dlerror(), we could encounter
a race with another thread.

In the DAOS control plane context, we always expect dlsym to
return a non-NULL value if it succeeded. This avoids detecting
an error set by another thread, when the current thread's dlsym()
succeeded.

Features: control

Signed-off-by: Kris Jacque <kristin.jacque@intel.com>