DAOS-12655 test: do not land, reproduction attempt
Log obj_ioc_init/fini() and references to thread-local
struct ds_cont_hdl objects. To investigate pool destroy hangs
due to held-references on struct ds_pool_child that may
be a result of IO not completed still holding these references.

Add a hack to ensure DEBUG log_mask is used in the engines when
running pool destroy in tearDown, which is where this intermittent
bug is observed.

Experiment with reducing /mnt/daos partition size (4->3GB) in this
particular test that only needs 1-2GiB of tmpfs. This is to save
memory for running engines in the memory-constrained VM environment.
i.e., to investigate a suspicion that laggy application performance
might be due to memory pressure (possibly starting to swap?) while
using tmpfs and running the daos_engines.

Skip-func-test-leap15: false
Quick-Functional: true
Test-tag: test_rebuild_with_io
Test-repeat: 10

Required-githooks: true

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>