DAOS-12412 test: Allow for log collection w/ No space left on device
Test log collection can fail if there is not enough space left on the
remote hosts to make a directory which is then used to clush --rcopy the
files to the launch host.  To avoid this ituation these directories will
be created prior to running the tests.

Skip-unit-tests:true
Test-tag-hw-large: dfs_test
Test-provider-hw-large: ucx+dc_x

Required-githooks: true

Signed-off-by: Phil Henderson <phillip.henderson@intel.com>