DAOS-12903 test: Allow for log collection w/ No space left on device (#11700)
Test log collection can fail if there is not enough space left on the
remote hosts to make a directory which is then used to clush --rcopy the
files to the launch host.  To avoid this ituation these directories will
be created prior to running the tests.

DAOS-12903 test: Fix path for pre-created archive directories (#11716)

Fix for commit 18c3c93.

Skip-unit-tests: true
Skip-func-hw-test: true

Signed-off-by: Phil Henderson <phillip.henderson@intel.com>