DAOS-12847 container: cache nhandles in batch close rdb tx
Before this change, when closing container handles from a pool handle
(e.g., as done for a crashed client, initiated by daos_agent), the
cont_close_hdls() ... get_nhandles() execution for a container
decremented its ds_cont_prop_nhandles and also read it back within
the same (uncommitted) rdb tx. This code violates the rule that an
rdb tx shall not read back its own updates. For containers with more
than one handle, this value was not decremented on close.

With this change, an rdb tx-scoped hash table is added to track by
container UUID, the number of open handles. get_nhandles() is updated
to, in this use case, search for the cached nhandles value in the cache
before updating the value in rdb. After an rdb tx is ended, the table
is destroyed and inserted records freed.

Also with this change the test_ior_crash functional test adds checks
of the container nhandles values after the test abruptly terminates ior.
The expected value is 1 to reflect that the ior open handles have been
closed as a result of daos_agent-based cleanup. The final ior run
(that gracefully finishes) is also checked to confirm nhandles == 1.

Test-tag: pr daily_regression,container full_regression,container test_ior_crash

Required-githooks: true

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>