DAOS-12843 ddb: Make sure VOS file is closed before exiting
The VOS file being connected to when running in non-interactive mode
was not being closed. By not closing the file, the system could
have difficulty starting again.  This change fixes that.

Also checks that if the user didn't close the file in interactive mode
it will be automatically closed.

Required-githooks: true

Signed-off-by: Ryon Jensen <ryon.jensen@intel.com>