DAOS-12841 snapshot: improve OIT scalability
Currently container snapshot Object ID tables are dumped
to a object with dkey 0 and each OID is stored as an akey,

so if there are a lot of OIDs in the container, snapshot
creation might easily get timeout. this patch tried to
improve problem by:

1. Use multiple dkeys/groups to store OIT table, RDB
is used to store OIT OID in case pool map changes.

2. add a new API daos_cont_snap_oit_dump() to dump
OIDs for a container snapshot, daos_cont_create_snap_opt()
is still kept to avoid break API.

Required-githooks: true
Signed-off-by: Wang Shilong <shilong.wang@intel.com>