DAOS-12189 chk: Include removed pools in query count (#11746)
When a pool has been removed, it will no longer appear
in the pools list for the response. We can still detect
that a pool was removed by checking the reports list,
however. If we find a removed pool, include it in the count
of checked pools when displaying the query results.

Includes misc. improvements to JSON output as well.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>