DAOS-12986 test: offline_parallel_test.py - Update kwargs to be passed into dmg_thread()
We need to pass in a dmg subcommand string (e.g., drain, exclude) to
dmg_thread() along with the necessary parameters and the results queue.

We used to pass in parameters for all actions (dmg subcommand string)
to dmg_thread() even though we only need to pass in parameters for a
particular action. This has caused an error in dmg_thread().

Update kwargs to store only the necessary parameters for a particular
action, along with the action string and the results queue.

Also make minor update to dmg_thread() to accompany with the change in
kwargs structure.

Skip-unit-tests: true
Skip-fault-injection-test: true
Test-tag: OSAOfflineParallelTest
Required-githooks: true

Signed-off-by: Makito Kano <makito.kano@intel.com>