DAOS-13035 bio: fix holes issue on replay
Unmap the unused region after repaly to erase the stale tx entries,
otherwise, stale tx could be mistakenly replayed on next restart in
the following scenario:

1. Imagine two inflight transactions T1 and T2, T1 is submitted before T2;
2. Before T1 is written to WAL, T2 is written successfully, both transactions
   are still regarded as incompleted since the preceding T1 is not persistent;
3. Sever restart;
4. WAL repaly hit the hole generated by unfinished T1 and stop relaying as
   expected, both T1 & T2 are not replayed;
5. A new transaction T3 committed, T3 happen to has same WAL size as T1, so it
   filled the hole perfectly;
6. Server restart again;
7. Both T3 & T2 are replayed since there is no way to tell that T2 is stale;

This unmap solves the issue for any device with unmap properly implemented,
but it won't be helpful for AIO device which doesn't support unmap. Given that
AIO device is only used for unit testing, and zeroing the unused region would
be too heavy, we choose to leave this risk for AIO device.

Required-githooks: true

Signed-off-by: Niu Yawei <yawei.niu@intel.com>