DAOS-13043 vos_tests: Abandon sleeps that used to be W/A DAOS-13040
Since PR-11828 is landed for DAOS-13043, the unit tests will be properly executed
and there is no need for extra sleep at the pool creation as a W/A for DAOS-13040.
The 'gen' collision probability is vanishingly small and the risk of the WAL test
false positive is negligible.

Required-githooks: true

Signed-off-by: Dmitry Ivanov <dmitry.ivanov2@hpe.com>