DAOS-13051 build: Use os.environ.copy() for SCONS_ENV option (#11853)
The env["ENV"] object must be implemented with some copy on write
or such because replacing it with os.environ means that environments
created with Clone() just get a reference to the environment.

I tested this with a small SConscript.  When replacing env["ENV"] with
os.environ wholesale, any changes in cloned environments appear in
the original environment as well.  With this change, the proper
behavior is restored.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>