DAOS-13074 test: target_failure.py - Set container property to healthy after reintegrate
PR11467 updated the container behavior so that a pool reintegrate
will not change the container status. The test was expecting the
container status to automatically go back to healthy after the
reintegrate. Update the test to manually set it to healthy.

PR11467 was cherry-picked to 2.2, but the change in target_failure.py
hasn't been included for some reason.

Added test tags.

Skip-unit-tests: true
Skip-fault-injection-test: true
Test-tag: TargetFailure
Required-githooks: true

Signed-off-by: Makito Kano <makito.kano@intel.com>