DAOS-0000 dfuse: Merge the info and projection_info structs.
There was duplication here from when we could support multiple
remotes so the distinction was artificial and there would only
ever be one of each struct, so merge the two.

Keep the old struct name as an alias to reduce the number of
lines touched by this cahnge, and heep the _iet entry with
the old prefix.

Test-tag: dfuse

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>