DAOS-623 test: Create python script for unit tests (#11869)
It's difficult to add new features and new tests that don't fit the standard in bash. This gives a little more flexibility.

Preemptively add placeholders for tests that are on feature branches so that post merge, they will still run.

Fix a bug in srv_checksum_test.c. It wasn't actually running under valgrind due to placement within run_test.sh, something that was unintentional and another reason to use python.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>