DAOS-5101 test: Update control/dmg_pool_evict test logic
The test purported to test handle eviction, but because it
was only using the daos tool, a new pool handle was being
created for each invocation of the tool. In order to properly
test eviction, the test needs to open a pool via the API and
hold the handle after eviction has occurred.

Features: control

Required-githooks: true

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>