DAOS-13074 test: Fix target_failure.py (#11877)
PR11467 updated the container behavior so that a pool reintegrate
will not change the container status. The test was expecting the
container status to automatically go back to healthy after the
reintegrate. Update the test to manually set it to healthy.

PR11467 was cherry-picked to 2.2, but the change in target_failure.py
hasn't been included for some reason.

Signed-off-by: Makito Kano <makito.kano@intel.com>