DAOS-13123 control: Increase pool timeout to 5 minutes
The current timeout of 3 minutes may be causing intermittent
failures in CI when a pool query happens during a rebuild.

Also updates the dRPC client implementation to return a
context.Canceled error when the dRPC connection is closed
due to timeout. This allows the harness to avoid incorrectly
invoking dRPC failure handlers for a canceled gRPC request.

Required-githooks: true

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>