DAOS-623 test: Only use open remote PR for base
I hit a situation where I resused a remote branch
from a PR that had been merged and the branch
deleted.  This caused my new PR to use a testbuild
branch for the base calculation.

Add a check to find_base.sh that the branch it
finds is an open PR.

Skip-build: true
Skip-test: true

Required-githooks: true

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>