DAOS-12153 cart: Correct mapping of HG_PROTOCOL_ERROR to DER_PROTO. (#11887)
The DAOS error code mapping of HG__PROTOCOL_ERROR (a generic, catch-all error) to DER_PROTO was put in place to resolve specific issue with a bulk handling error returned from Mercury. The mapping allowed the bulk failure to be handled correctly, since DER_PROTO is not a retry-able error. The change to mapping has resulted in the incorrect reporting of “Incompatible Protocol” (the intended meaning of DER_PROTO) for some network errors. The fix for this in linked PR is to remap HG_PROTOCOL_ERROR to a newly-added error code: DER_HG_FATAL, meaning a non-retry-able Mercury error has occurred.

Note that this fix will not fully resolve the potential overuse of DER_PROTO in code. It is used in security (credential errors), object (srv_obj.c for several conditions), and a few other modules. Note that although the text for the error message (and the comment with the declaration) state the error as “Incompatible Protocol” this error is not returned on incompatibility by the CaRT protocol_query RPC, which instead returns -DER_MISMATCH. A related PR will eliminate the separate protocol query RPC, and will thus will address any ambiguity for that functionality. My suggestion for further improvement is to either define “incompatibility” properly, or apply more specific error codes to the various places in the code where it is currently used.

Signed-off-by: Joseph Moore <joseph.moore@intel.com>