DAOS-13114 control: Add batching support for PoolEvict requests (#11893)
The PoolEvict gRPC handler may be invoked from N agents in the
event of a large-scale application crash. This would result in
N dRPC requests from the MS leader to its local engine in order
to evict the pool handles in those requests. With this commit,
a batching mechanism is added to improve performance in these
large-scale events so that N agent requests are consolidated
into a smaller number of dRPC requests containing multiple
handles.

Also improves agent logging of evicted handles due to application
crash by including the process name and logging a single line
at INFO level instead of DEBUG.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>