DAOS-12435 control: Remove dmg storage query target-health cmd
For MD-on-SSD target information and data can be spread across
multiple SSDs on multiple tiers, as such the output of dmg storage
query target-health command may be less useful because it will include
multiple SSDs including those that either contain meta data or data
for the given target. Since this command was mainly added for
development purposes and seems not to be widely used, remove the
command to reduce unnecessary code and command surface.

Required-githooks: true

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>