DAOS-9755 control: Add DD_MASK opt to dmg set-logmasks
DD_MASK envariable can be used to specify which debug streams to
enable when logging on the server. This allows for finer grained log
filtering and this change adds support to supply an extra optional
argument to dmg server set-logmasks which would update the engine
process to use the new setting on relevant storage servers.

Required-githooks: true

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>