DAOS-13170 chk: dmg check enable should not restart ranks (#12030)
In order to avoid unexpected behavior, the `dmg check enable`
command now requires the system to be stopped first. Likewise,
the `dmg check disable` command just disables the checker
and stops the ranks. The admin will need to use `dmg system start`
again in order to restart the ranks in normal mode.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>