DAOS-13194 common: prealloc bunch of environment vars slots
Since late setenv()s are racy with getenv()s, and can occur in
DAOS and pre-reqs/3rd-parties, try to preallocate a bunch of
environment vars slots by calling setenv() hundreds of times
for dummy env vars, and then unsetenv() them all in a raw.
Doing so will allow for the environ vars pointers array not
to be re-allocated and thus prevent crashes.

Required-githooks: true

Signed-off-by: Bruno Faccini <bruno.faccini@intel.com>