DAOS-13448 cart: Reject stale incoming PG versions
It has been observed that the MS leader may pass system map {rank 1,
rank 2} with version 2 to the local rank 0, who may mistakenly think
that itself has been excluded from the system (it actually has not even
joined the system in the first place). Similarly, if a rank joins in
system map version V, and somehow receives a system map version V - 1
that doesn't include itself, it may also mistakenly think that itself
has been excluded from the system.

To address these two scenarios, we pass a minimum group version---the
system map version in which a rank joins the system---from the MS to the
crt PG on this rank via JoinResp, so that the rank will reject any
incoming PG version that is less than the minimum group version. This
also implements the to-do in crt_swim_cli_cb.

Signed-off-by: Li Wei <wei.g.li@intel.com>
Required-githooks: true