DAOS-13227 dfuse: Use 16 threads by default.
Default to 16 threads rather than 1-per-core.
If a cpuset is defined then use 1 thread per core still,
however if nothing is defined then limit to 16.

It is still possible to over-commit by using the command line.

Switch to using hwloc to read the cpusets as the standard Linux
functions only tell you the allowed size, not the total
size.

Required-githooks: true

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>