DAOS-13515 vos: use vos format to store SMD
Now for MD-on-SSD, we store SMD on the external path with
LMMDB which ntroduced some extra work for Catastrophic recovery work. extra tool is needed
to parse and verify LMMDB. It will be nice to unify to use VOS format, idea is we still
could store SMD on the external path but rather than use LMMDB, VOS format is still kept,
this might be ok since performance is not sensitive for SMD operations.

Required-githooks: true
Signed-off-by: Wang Shilong <shilong.wang@intel.com>