SRS-394 common: Metadata on SSD Phase 1 (#12228)
Adds Metadata on SSD feature to DAOS.

For Phase 1, all metadata is stored in DRAM. The metadata blob on SSD
will be logically an exact replica of the DRAM. In order to avoid too frequent,
byte granular updates to the SSD, we instead commit enough information to replay
updates to the write ahead log (WAL). A background checkpointing process runs
periodically to sync the dram pool with the SSD metadata blob and to reclaim
space in the WAL. The initial allocator of space within the metadata blob (DAV) is
based on the PMDK allocator. This patch also contains various control plane changes
necessary to support both PMDK and MD on SSD modes of operation. SMD replaces
VOS with lmdb in this version but otherwise, the fundamental changes are at or below
the VOS layer.

Signed-off-by: Liang Zhen <liang.zhen@intel.com>
Signed-off-by: Wang Shilong <shilong.wang@intel.com>
Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>
Signed-off-by: Niu Yawei <yawei.niu@intel.com>
Signed-off-by: Xuezhao Liu <xuezhao.liu@intel.com>
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Signed-off-by: ylivis <yann.livis@hpe.com>
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Signed-off-by: Kris Jacque <kristin.jacque@intel.com>
Signed-off-by: Dmitry Ivanov <dmitry.ivanov2@hpe.com>
Signed-off-by: Ted Anderson <nonzero@google.com>
Signed-off-by: Phil Henderson <phillip.henderson@intel.com>
Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>
Signed-off-by: Bruno Faccini <bruno.faccini@intel.com>
Signed-off-by: Lei Huang <lei.huang@intel.com>
Signed-off-by: onever23 <oleg.neverovitch@hpe.com>
Signed-off-by: Sherin T George <sherin-t.george@hpe.com>
Signed-off-by: Sydney Vanda <sydney.m.vanda@intel.com>