DAOS-8739 control: Show more detailed results from dmg set-logmasks (#12240)
Prior to this change, only the failures would be reported and it would
not be obvious to the user as to which hosts succeeded in applying the
requested log-masks. Results are now reported in both an error table
(printed to stderr) indicating which hostset failed with a given error
string and a hostset signifying successful operations (printed to
stdout). Each string printed in the error table will indicate the
result of each engine on the host.

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>