DAOS-12751 tool: Add a daos filesystem evict command.
Add a command to tell dfuse and the kernel to evict a
path in dfuse.  This is implemented via a ioctl that
sets a flag on the file handle, when the file is then
closed the kernel is then told the dentry for the file
is no longer valid which should cause the kernel to
forget the entry and any contents if the entry is
a directory.

Required-githooks: true

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>