DAOS-12250 control: Log gRPC req before processing (#12345)
The previous commit placed the request log behind a
conditional check in order to try to cut down on noisy
logging for requests that often result in a sentinel error
(e.g. ErrNotLeader). That approach does result in cleaner
logs, but has the downside of making requests more difficult
to trace, as the beginning of the request handling is not
logged first.

This commit adjusts the logging interceptor to restore the
previous behavior of logging the request before handling it,
but first checks to see if the request should be logged based
on the server's leadership state. The end result should be that
Pool requests in particular are only logged by the server that
handles them.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>