DAOS-13919 control: Fix dmg pool get-prop compat with v2.2 pools (#12708)
JSON marshalling fails when attempting to print PoolGetProps()
control-API response as some of the returned properties have no
associated value set. This can occur when dmg pool get-prop is called
from a >= v2.4 dmg/daos_server using a pool that was created with v2.2.
This occurs because the old pool has a reduced set of properties and so
those that don't have an associated value in the RPC response end up
as PoolProperty objects with no associated value (which is not allowed
in the JSON marshal implementation for the type).

Solution is to remove the missing properties from the response before
sending to dmg, the control-API consumer. This does mean that
get-prop calls can return a different property list depending on
which DAOS version was used to create the pool.

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>