DAOS-13988 control: MemberResult state should reflect error (#12719) (#12754)
Change addresses the issue where dmg system stop fails if some ranks
have already been previously stopped. This is due to a couple of bugs.
First a situation exists where the state conveyed in a MemberResult
doesn't match the fact that the result was a failure. The consumer of
the results checks this and raises an error in the case that they don't
match. Secondly the PrepShutdown flow which is triggered as stage-one
of dmg system stop returns a failed result for an already stopped rank.

Return successful results when calling PrepShutdown on stopped ranks
Ensure errored MemberResults convey a matching error state

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>