DAOS-13814 csum: Copy iod size for checksums
The iod size is required for checksums while they are being gathered
on an object fetch. In some situations the iods structure that the
checksums are using is different than the iods structure that is used
during a vos object fetch. This change copies the iod sizes if needed.

Features: checksum
Required-githooks: true

Signed-off-by: Ryon Jensen <ryon.jensen@intel.com>