DAOS-14119 object: fix a bug of shards_scheded setting (#12783)
In obj_shard_task_sched() it schedules the shard tasks and then
set "obj_auxi->shards_scheded = 1". It is not safe as once shard
tasks scheduled it possibly be executed immediately by other
threads and completed, that cause the obj_auxi be invalid and cause
mem corruption. The problem found by Mohamad.

Signed-off-by: Xuezhao Liu <xuezhao.liu@intel.com>