DAOS-14139 test: Debug Unhandled exception for dfuse/caching_check.py
During the teardown of dfuse/caching_checking.py, Unhandled exception occurs
when dmg pool destroy is called. The command itself works, but we can't
tell where that exception is coming from, so use Exception in remove_pool()
so that we can catch and investigate any kind of exception.

This issue is intermittent.

Skip-unit-tests: true
Skip-fault-injection-test: true
Test-tag: test_dfuse_caching_check
Test-repeat: 6
Required-githooks: true

Signed-off-by: Makito Kano <makito.kano@intel.com>