DAOS-14085 control: Fix config generate yaml printed to stdout (#12770) (#12845)
INFO level log messages were being printed to stdout during the
execution of dmg config generate meaning that the output could not be
redirected to a .yml file and used as a daos_server config. Fix by
adjusting the effective log level to ERROR during execution of the
ConfigGenerate Control-API call. The INFO and NOTICE logging can
safely be ignored as these messages are intended for daos_server users
when running in daemon mode. In order to preserve --debug cmdline
functionality, don't adjust the log level if running with the flag set.

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>