DAOS-14216 build: Build fail if home is under /usr
The DAOS build will fail if a user builds prerequisites
under /usr such as when the home directory is nested
within.  This patch simply changes the check for
whether to patch rpath to continue if prereq is in
the path for the binary in question.

Also, patch the clang version check to let custom build
work.

Required-githooks: true

Signed-off-by: Jeff Olivier <jeffolivier@google.com>