DAOS-14139 test: Abort testing if there's avocado timeout in _cleanup() of test.py
In _cleanup(), cleanup methods are called from the cleanup method stack.
If the method causes an exception, we'll add the error to the list and
move on to the next method. However, if the exception is a timeout, it would
print unhandled exception in the log and it would be hard to debug. Thus,
if the exception is timeout, abort testing.

Skip-unit-tests: true
Skip-fault-injection-test: true
Test-tag: test_dfuse_caching_check
Test-repeat: 12
Required-githooks: true
Signed-off-by: Makito Kano <makito.kano@intel.com>