DAOS-14332 control: Use gRPC metadata for interop
The interoperability checking code relies on getting the
peer component from the peer certificate. When running
in insecure mode, the peer component information is
unavailable, and the interoperability check defaults
to the most stringent requirements.

This patch adds new component/version headers to the
grpc client request to allow the server to perform
interoperability checks without the peer certificate.

Features: control

Required-githooks: true
Change-Id: I1d9a6d171ecadded3bf041a36e2a156881d93f57
Signed-off-by: Michael MacDonald <mjmac@google.com>