DAOS-14105 object: collective punch object
Currently, when punch an object with multiple redundancy groups,
to guarantee the atomicity, we handle the whole punch via single
internal distributed transaction. The DTX leader will forward the
CPD RPC to every object shard within the same transaction. For a
large-scaled object, such as a SX object, punching it will generate
N RPCs (N is equal to the count of all the vos targets in the system).
That will be very slow and hold a lot of system resource for relative
long time. If the system is under heavy load, related RPC(s) may get
timeout, then trigger DTX abort, and then client will resend RPC to
the DTX leader for retry, that will make the situation to be worse
and worse.

To resolve such bad situation, we will introduce new mechanism:
collective punch the object. The basic idea is that: when punch
an object with multiple redundancy groups, the client will send
OBJ_COLL_PUNCH RPC to the DTX leader. On the DTX leader, instead
of forwarding the request to all related vos targets, it forwards
the request only to all involved engines. And then related engines
will generate collective tasks to punch the object shards on each
own local vos targets. That will save a lot of RPCs and resources.

Signed-off-by: Fan Yong <fan.yong@intel.com>