DAOS-12541 ddb: Improved List Output (#12300)
* DAOS-12541 ddb: Improved List Output

Replaced the "vos path builder" with "indexed tree path" structure.
It is used for parsing and printing vos paths. The 'ls' command
is updated to print the paths now to make it possible to copy/paste
the paths. The indexed_tree_path better supports dkey and key values
by using a syntax for specifying type (example: '{uint32: 0x123abc}')
and length of a string (example: 'akey{5}'). Can also escape special
characters using the '\' in keys.

Signed-off-by: Ryon Jensen <ryon.jensen@intel.com>