DAOS-14352 object: right epoch to fetch the data
If the epoch is higher than EC aggregate boundary, then
it should use stable epoch to fetch the data, since
the data could be aggregated independently on parity
and data shard, so using stable epoch could make sure
the consistency view during rebuild. And also EC aggregation
should already aggregate the parity, so there should not
be any partial update on the parity as well.

Otherwise there might be partial update on this rebuilding
shard, so let's use the epoch from the parity shard to fetch
the data here, which will make sure partial update will not
be fetched here. And also EC aggregation is being disabled
at the moment, so there should not be any vos aggregation
impact this process as well.

Features: rebuild ec
Required-githooks: true

Signed-off-by: Di Wang <di.wang@intel.com>