DAOS-14010 container: maintain stable epoch
Maintain stable epoch on each target as the earliest
open dtx epoch - 1, which will be reported to the PS leader,
which will choose the minimum one from all target, then
sync it to all targets as the VOS aggregation upper
boundary, and also as the stable epoch of each target, which
will be used by incremental reintegration as the start
epoch to reconstruct the data.

Add daos container version in vos_cont_df.
Required-githooks: true

Signed-off-by: Di Wang <di.wang@intel.com>