DAOS-14270 common: Workaround setenv reallocation
For interception libraries, deferring calls to daos_init
can cause issues because DAOS calls setenv.  If the
application is threaded at this point, this can cause a
crash because the environment is global and may get
reallocated. This patch enables a workaround where
daos can pre-set environmentment variables to a
known "empty" value.  The wrappers treat this value
as empty.  This enables the interception libraries
to preallocate environment variables they may set
without causing a reallocation when they are set
at runtime.

Required-githooks: true

Signed-off-by: Jeff Olivier <jeffolivier@google.com>