DAOS-0000 dfuse: Avoid using decref on inods in dfuse.
When the kernel performs a fuse operations it already holds
a reference on the inode, so it is not required for dfuse to
take a second one.  DFuse needs to perform a lookup which
will take a reference but rather than holding this reference
until after replying to the kernel and then calling
decref which will take locks etc simply decrease the ref count
after the lookup using atomics.

This is safe as the reference being dropped is never the last
one so locks are not needed.

Assert that the inode is found for fuse operations, if it isn't
then something serious has gone wrong so assert is the correct
check here.

Overall this decreases hash table operations considerably and
removes both the need for locking and also code-paths where
inodes could potentially be closed.

Test-tag: dfuse
Required-githooks: true

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>